pyproject.toml
[project]
name = "pedaldrivenprogramming"
version = "3.0.0"
dependencies = [
    "Django",
    "dj-database-url",
    "markdown",
    "psycopg",
    "python-frontmatter",
    "whitenoise[brotli]",
]

[project.optional-dependencies]
dev = [
    "django-stubs",
    "pip-tools",
    "ruff",
    "watchfiles"
]

[tool.setuptools]
py-modules = []
